pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=1.1.7",
    "matplotlib>=3.10.8",
    "mcp-use>=1.6.0",
    "torch",             # Added explicitly so we can manage its source below
    "llamafirewall",     # Added your target library
]

# 1. Define the custom repository for CPU-only wheels
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

# 2. Force 'torch' to ALWAYS come from that CPU repository
[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }  # Added just in case dependencies ask for it